requirements.txt
pandas>=2.2
numpy>=1.26
pyarrow>=15
python-dateutil>=2.9
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.1
pyyaml>=6.0
scikit-learn>=1.4
xgboost>=2.0
matplotlib>=3.8
openpyxl>=3.1
